code fix for NP_NULL_ON_SOME_PATH #4273
